requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# OpenAI
openai==1.54.0
httpx==0.27.2

# LangChain and RAG
langchain==0.1.6
langchain-community==0.0.19
chromadb==0.4.22

# Document loaders
pypdf==4.0.1
unstructured[local-inference]==0.12.2
openpyxl==3.1.2
python-magic==0.4.27
python-magic-bin==0.4.14; sys_platform == 'win32'

# Embeddings
sentence-transformers==2.3.1
torch==2.2.0
transformers==4.37.2

# Utilities
pydantic==2.6.0
python-dotenv==1.0.1
aiofiles==23.2.1

